Initial commit.
Start fleshing out a platform for parsing KiCad .sch files, with a focus
on mangling components contained therein.

Currently accepts a file dragged onto a drop zone and reads it in-place,
rending a sorted table of components found in the file.

Behind the scenes, it is parsed into objects, with a lot of information
on components pulled out of the file in a structured way.